phutil: Persistence Homology Utilities

A low-level package for hosting persistence data. It is part of the 'TDAverse' suite of packages, which is designed to provide a collection of packages for enabling machine learning and data science tasks using persistent homology. Implements a class for hosting persistence data, a number of coercers from and to already existing and used data structures from other packages and functions to compute distances between persistence diagrams. A formal definition and study of bottleneck and Wasserstein distances can be found in Bubenik, Scott and Stanley (2023) doi:10.1007/s41468-022-00103-8 . Their implementation in 'phutil' relies on the 'C++' Hera library developed by Kerber, Morozov and Nigmetov (2017) doi:10.1145/3064175 .

See also

Author

Maintainer: Aymeric Stamm aymeric.stamm@cnrs.fr (ORCID)

Authors:

Jason Cory Brunson cornelioid@gmail.com (ORCID)

Other contributors:

Michael Kerber (HERA C++ code) [contributor]

Dmitriy Morozov (HERA C++ code) [contributor]

Arnur Nigmetov (HERA C++ code) [contributor]